refactor(opencrud-schema-provider): convert to typescript
