+++ packages/opencrud-schema-provider/.eslintignore
@@ -1,2 +1,3 @@
 node_modules
 coverage
+lib
