+++ packages/opencrud-schema-provider/modules.d.ts
@@ -1,0 +1,2 @@
+declare module '@venncity/sequelize';
+declare module '@venncity/clou-utils';
