--- packages/opencrud-schema-provider/src/index.js
@@ -1,17 +1,0 @@
-const {
-  openCrudDataModel,
-  openCrudSchema,
-  openCrudSchemaGraphql,
-  getOpenCrudIntrospection
-} = require('./openCRUDSchemaProvider/openCRUDSchemaProvider');
-const graphqlSchemaUtils = require('./schemaUtils/graphqlSchemaUtils');
-const introspectionUtils = require('./schemaUtils/introspectionUtils');
-
-module.exports = {
-  openCrudDataModel,
-  openCrudSchema,
-  openCrudSchemaGraphql,
-  getOpenCrudIntrospection,
-  graphqlSchemaUtils,
-  introspectionUtils
-};
