+++ packages/opencrud-schema-provider/src/index.ts
@@ -1,0 +1,12 @@
+import { openCrudDataModel, openCrudSchema, openCrudSchemaGraphql, getOpenCrudIntrospection } from './openCRUDSchemaProvider/openCRUDSchemaProvider';
+import graphqlSchemaUtils from './schemaUtils/graphqlSchemaUtils';
+import introspectionUtils from './schemaUtils/introspectionUtils';
+
+export default {
+  openCrudDataModel,
+  openCrudSchema,
+  openCrudSchemaGraphql,
+  getOpenCrudIntrospection,
+  graphqlSchemaUtils,
+  introspectionUtils
+};
