+++ packages/opencrud-schema-provider/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "lib/cjs",
+    "noEmit": false,
+    "noImplicitAny": false
+  },
+  "include": [
+    "src",
+    "modules.d.ts"
+  ],
+  "exclude": [
+    "node_modules",
+    "lib",
+    "coverage"
+  ]
+}
